Extract process-metric calculators from process_control.py
First thematic extraction from modules/process_control.py:

- modules/process_metrics.py (~165 lines, new):
  4 functions + 3 module-level caches for computing CPU% and disk-IO
  stats over the Kometa/ImageMaid subprocess trees.
  * calculate_process_cpu_percent(proc)
  * calculate_system_cpu_percent()
  * calculate_process_io_stats(proc, cache_name)
  * clear_process_metric_cache(pid, cache_name=None)
  * Module state: KOMETA_CPU_CACHE, SYSTEM_CPU_CACHE, PROCESS_IO_CACHE

modules/process_control.py: 1259 -> 1155 (-104 / -8.3%).

Small extraction to establish the pattern for the process_control.py
split.  This cluster was chosen first because:
1. It's completely self-contained (no dependencies on other pc functions)
2. Its module-level caches are truly private (zero external refs)
3. It only depends on stdlib + psutil (no cross-cluster imports)
4. Tests already patch these via qs_module.__dict__ aliases, so
   monkeypatch continues working across the module boundary

process_control.py re-exports every moved name so external callers
(quickstart.py imports all 4 via alias in its big process_control
import block; modules/logscan_*.py also import a few) keep working
unchanged.

AST-verified: all 4 functions are byte-for-byte identical to
origin/develop.
